bot/requirements.txt
aiogram==3.13.1
anthropic==0.42.0
aiohttp==3.11.11
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
aiofiles==24.1.0
